azure_pipeline/src/requirements.txt
azure-functions>=1.14.0
azure-cosmos>=4.3.0
azure-storage-blob>=12.14.0
azure-identity>=1.12.0
python-dateutil>=2.8.2